requirements.txt
Flask==2.2.3
flask-cors==3.0.10
RPi.GPIO==0.7.1
Adafruit_DHT==1.4.0
PyYAML==6.0